Update patch set 4
Patch Set 4:

Steve,

personally I would _not_ ifdef it. I asked around among the wl12xx maintainers, and at least Pavel (for WL1271 on Milestone) and Ricardo (for WL1283 on LG P99x) told me that seemingly the WLAN indicator is stuck for them as well.

I tried to understand how this is supposed to work for quite a long time, but I failed so far. From my understanding, the current implementation for the rssi-approx command returns the RSSI value cached from the last scan run. What I don't understand is how this cache is supposed to be updated when connected to an AP, as no scan runs in that case. The rssi command implementation always asks the driver, which seems the right thing to do to me in any case, which is why I made them do the same thing.

I've included the WL12xx device maintainers I'm aware of, maybe they can comment as well.

All,

does the WLAN signal strength indicator work properly on your devices when connected to an AP and changing the reception, e.g. by moving farther away from the AP?

Patch-set: 4