Update patch set 1
Patch Set 1:

the commit message is not precise, my bad, sorry. we use some patches from here:
http://omapzoom.org/?p=platform/hardware/ti/wlan.git;a=shortlog;h=refs/heads/p-jb-release 

we will get back to the jb release and test if these changes are needed or not. This one works fine with wifi tethering. I need a clean test yet.
 
Anyone having trouble with wifi tethering on ti-wlan and jelly bean without this patch?

Patch-set: 1