Update patch set 1
Patch Set 1:

I didn't have time to test these patches on cm-10.2

On cm-10.1 they seem to work fine but on 10.2 I have problems with hotspot too

I mainly pulled these patches to test the hotspot behaviour, because it kept broadcasting even when it was turned off and it was draining the battery until a reboot.

Maybe we should just wait for TI's 4.3 updates.

Patch-set: 1