Update patch set 5
Patch Set 5:

Our problem with encore stems from a symlink that we need to use to export our mac address. Our address is found in an obscure configuration file which we read and then insert into the wl1271-nvs_127x.bin to create a new (912 byte) wl1271-nvs.bin file which resides in data so it will survive flashes. This is symlinked to /system/etc/firmware/ti-connectivity/wl1271-nvs.bin I will look at ways to get around this. There ought to be a better way.

Patch-set: 5