Update patch set 1
Patch Set 1:

With all the patches in the "backports" topic (except one that is HTC specific) I have a fully working Wi-Fi driver. At least that's my impression so far. AP and P2P are are working and the BOARD_WIFI_SKIP_CAPABILITIES flag is also not needed.

Arik Nemtsov told me that they test their TI supplicant with the backports driver. So while at the moment it's a fairly invasive change, it may help avoiding incompatibilities in the future.

I haven't yet looked at the driver calibration.

Patch-set: 1