Update patch set 1
Patch Set 1:

I'd prefer to have this commited. This time it's even a "official" commit and not my hackery so we can assume that it is an okay thing to do.
You're not left alone wondering whether or not to define STUB is the right fix for your compilation problem.

Patch-set: 1
Reviewer: Thomas Wendt <5678@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0