wlcore/wl18xx: fw logger over sdio
enable the FW Logger to work over the SDIO interface instead of only
over UART.
in new design we will use fw internal memory instead of packet ram
that was used in older (and wl12xx) design.
this change will reduce the impact on tp and stability.
adding new event to notify fw logger is ready to be read.
adding dynamic configuration to debugfs.

Signed-off-by: Shahar Patury <shaharp@ti.com>
Change-Id: I52923a172e2e8d22e9c493ff1c5df79bec0e0800
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>