Update patch set 1
Patch Set 1: Code-Review+1

This was the last piece of the puzzle; WoWLAN needs to be setup from userspace to be able to keep Wi-Fi always on. For K3.0, the top changes in https://github.com/airend/android_kernel_bn_omap/commits/stable (18 or so) are mostly required to align to new pdata.

Patch-set: 1
Label: Code-Review=+1