Update patch set 1
Patch Set 1:

> What's the status of this patch series? Are there any big issues
 > left or I should go ahead and try it?

The one mild issue left is that power consumption with WoWLAN enabled is higher than expected. W/o WoWLAN, it's great (a month estimated standby, vs a week), but you also lose constant connectivity. The cfg80211_leave_all() change happened with K3.10, and this compat stack is based on 4.1+, so I don't expect things to be very different in R8.7… I'll test a few more things, including the last fw-4 with WoWLAN triggers, but the current changes are definitely usable/good for testing.

Patch-set: 1