Update patch set 1
Patch Set 1:

BTW, I meant to add that pattern filtering is not really needed after all. Most of the battery drain in ELP was unrelated, and will push updated changes shortly. I will leave this commit outside the series in case we want to look at it again later.

Patch-set: 1